+++ test/widget_test.mocks.dart
@@ -33,6 +33,15 @@ class MockAuthService extends _i1.Mock implements _i2.AuthService {
   }
 
   @override
+  set onSessionInvalidated(void Function()? listener) => super.noSuchMethod(
+        Invocation.setter(
+          #onSessionInvalidated,
+          listener,
+        ),
+        returnValueForMissingStub: null,
+      );
+
+  @override
   _i3.Future<void> signInWithGoogle() => (super.noSuchMethod(
         Invocation.method(
           #signInWithGoogle,
@@ -52,6 +61,25 @@ class MockAuthService extends _i1.Mock implements _i2.AuthService {
         returnValue:
             _i3.Future<Map<String, dynamic>>.value(<String, dynamic>{}),
       ) as _i3.Future<Map<String, dynamic>>);
+
+  @override
+  _i3.Future<Map<String, dynamic>?> readCachedUser() => (super.noSuchMethod(
+        Invocation.method(
+          #readCachedUser,
+          [],
+        ),
+        returnValue: _i3.Future<Map<String, dynamic>?>.value(),
+      ) as _i3.Future<Map<String, dynamic>?>);
+
+  @override
+  _i3.Future<void> invalidateSession() => (super.noSuchMethod(
+        Invocation.method(
+          #invalidateSession,
+          [],
+        ),
+        returnValue: _i3.Future<void>.value(),
+        returnValueForMissingStub: _i3.Future<void>.value(),
+      ) as _i3.Future<void>);
 
   @override
   _i3.Future<String> getValidAccessToken() => (super.noSuchMethod(